Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   .vscode/settings.json
	deleted:    Luogu/string/3.calc.cpp
	new file:   Luogu/string/4.calc.cpp
	new file:   Luogu/string/4.calc.exe
	new file:   Luogu/string/5.countTitle.cpp
	new file:   Luogu/string/5.countTitle.exe
	new file:   Luogu/string/6.wordProcess.cpp
	new file:   Luogu/string/6.wordProcess.exe
	new file:   Luogu/string/7.wordCount.cpp
	new file:   Luogu/string/7.wordCount.exe
	modified:   experiment/2/202culExpression.exe
	modified:   experiment/2/203triangle.cpp
	modified:   experiment/2/203triangle.exe
	modified:   experiment/2/204calculater.cpp
	modified:   experiment/2/204calculater.exe
	modified:   experiment/2/205countCharacter.cpp
	modified:   experiment/2/205countCharacter.exe
	modified:   experiment/2/206GCD&LCM.cpp
	modified:   experiment/2/206GCD&LCM.exe
	modified:   experiment/2/207output.exe
	modified:   experiment/2/208sqrt.exe
	modified:   experiment/2/209apple.exe
	new file:   practice/lianbiao.c
	new file:   practice/lianbiao.exe
	modified:   practice/test.cpp
	modified:   practice/test.exe
	modified:   practice/text.cpp
	modified:   practice/text.exe

+++ experiment/2/204calculater.cpp
@@ -20,7 +20,7 @@ int main()
 	}
 	else if (op == '/' && num2 != 0)
 	{
-		cout << num1 << '/' << num2 << '=' << num1 / num2;	
+		cout << num1 << '/' << num2 << '=' << double(num1) / num2;	
 	}
 	else if(op == '%')
 	{
